+++ tests/testthat/test-embedding.R
@@ -7,7 +7,12 @@ test_that("embed_adjacency_matrix -- Undirected, unweighted case works", {
   no <- 7
   A <- g[]
   A <- A +
-    1 / 2 * as(Matrix::Matrix(diag(degree(g)), doDiag = FALSE), "generalMatrix")
+    1 /
+      2 *
+      as(
+        Matrix::Matrix(diag(degree(g, mode = "all")), doDiag = FALSE),
+        "generalMatrix"
+      )
   ss <- eigen(A)
 
   U <- standardize_eigen_signs(ss$vectors)
@@ -17,14 +22,14 @@ test_that("embed_adjacency_matrix -- Undirected, unweighted case works", {
     g,
     no = no,
     which = "la",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = TRUE
   )
   as_la <- embed_adjacency_matrix(
     g,
     no = no,
     which = "la",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = FALSE
   )
 
@@ -40,14 +45,14 @@ test_that("embed_adjacency_matrix -- Undirected, unweighted case works", {
     g,
     no = no,
     which = "lm",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = TRUE
   )
   as_lm <- embed_adjacency_matrix(
     g,
     no = no,
     which = "lm",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = FALSE
   )
 
@@ -66,14 +71,14 @@ test_that("embed_adjacency_matrix -- Undirected, unweighted case works", {
     g,
     no = no,
     which = "sa",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = TRUE
   )
   as_sa <- embed_adjacency_matrix(
     g,
     no = no,
     which = "sa",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = FALSE
   )
 
@@ -94,7 +99,12 @@ test_that("embed_adjacency_matrix -- Undirected, weighted case works", {
   no <- 3
   A <- g[]
   A <- A +
-    1 / 2 * as(Matrix::Matrix(diag(degree(g)), doDiag = FALSE), "generalMatrix")
+    1 /
+      2 *
+      as(
+        Matrix::Matrix(diag(degree(g, mode = "all")), doDiag = FALSE),
+        "generalMatrix"
+      )
   ss <- eigen(A)
 
   U <- standardize_eigen_signs(ss$vectors)
@@ -104,14 +114,14 @@ test_that("embed_adjacency_matrix -- Undirected, weighted case works", {
     g,
     no = no,
     which = "la",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = TRUE
   )
   as_la <- embed_adjacency_matrix(
     g,
     no = no,
     which = "la",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = FALSE
   )
 
@@ -127,14 +137,14 @@ test_that("embed_adjacency_matrix -- Undirected, weighted case works", {
     g,
     no = no,
     which = "lm",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = TRUE
   )
   as_lm <- embed_adjacency_matrix(
     g,
     no = no,
     which = "lm",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = FALSE
   )
 
@@ -153,14 +163,14 @@ test_that("embed_adjacency_matrix -- Undirected, weighted case works", {
     g,
     no = no,
     which = "sa",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = TRUE
   )
   as_sa <- embed_adjacency_matrix(
     g,
     no = no,
     which = "sa",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = FALSE
   )
 
@@ -178,7 +188,12 @@ test_that("embed_adjacency_matrix -- Directed, unweighted case works", {
   no <- 3
   A <- g[]
   A <- A +
-    1 / 2 * as(Matrix::Matrix(diag(degree(g)), doDiag = FALSE), "generalMatrix")
+    1 /
+      2 *
+      as(
+        Matrix::Matrix(diag(degree(g, mode = "all")), doDiag = FALSE),
+        "generalMatrix"
+      )
   ss <- svd(A)
 
   U <- standardize_eigen_signs(ss$u)
@@ -190,14 +205,14 @@ test_that("embed_adjacency_matrix -- Directed, unweighted case works", {
     g,
     no = no,
     which = "la",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = TRUE
   )
   as_la <- embed_adjacency_matrix(
     g,
     no = no,
     which = "la",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = FALSE
   )
 
@@ -218,14 +233,14 @@ test_that("embed_adjacency_matrix -- Directed, unweighted case works", {
     g,
     no = no,
     which = "lm",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = TRUE
   )
   as_lm <- embed_adjacency_matrix(
     g,
     no = no,
     which = "lm",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = FALSE
   )
 
@@ -246,14 +261,14 @@ test_that("embed_adjacency_matrix -- Directed, unweighted case works", {
     g,
     no = no,
     which = "sa",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = TRUE
   )
   as_sa <- embed_adjacency_matrix(
     g,
     no = no,
     which = "sa",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = FALSE
   )
 
@@ -279,7 +294,12 @@ test_that("embed_adjacency_matrix -- Directed, weighted case works", {
   no <- 3
   A <- g[]
   A <- A +
-    1 / 2 * as(Matrix::Matrix(diag(degree(g)), doDiag = FALSE), "generalMatrix")
+    1 /
+      2 *
+      as(
+        Matrix::Matrix(diag(degree(g, mode = "all")), doDiag = FALSE),
+        "generalMatrix"
+      )
   ss <- svd(A)
 
   U <- standardize_eigen_signs(ss$u)
@@ -291,14 +311,14 @@ test_that("embed_adjacency_matrix -- Directed, weighted case works", {
     g,
     no = no,
     which = "la",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = TRUE
   )
   as_la <- embed_adjacency_matrix(
     g,
     no = no,
     which = "la",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = FALSE
   )
 
@@ -317,14 +337,14 @@ test_that("embed_adjacency_matrix -- Directed, weighted case works", {
     g,
     no = no,
     which = "lm",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = TRUE
   )
   as_lm <- embed_adjacency_matrix(
     g,
     no = no,
     which = "lm",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = FALSE
   )
 
@@ -343,14 +363,14 @@ test_that("embed_adjacency_matrix -- Directed, weighted case works", {
     g,
     no = no,
     which = "sa",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = TRUE
   )
   as_sa <- embed_adjacency_matrix(
     g,
     no = no,
     which = "sa",
-    cvec = degree(g) / 2,
+    cvec = degree(g, mode = "all") / 2,
     scaled = FALSE
   )
 
@@ -398,7 +418,10 @@ test_that("embed_laplacian_matrix -- Undirected, unweighted, D-A case works", {
   g <- sample_gnm(10, 20, directed = FALSE)
 
   no <- 3
-  A <- as(Matrix::Matrix(diag(degree(g)), doDiag = FALSE), "generalMatrix") -
+  A <- as(
+    Matrix::Matrix(diag(degree(g, mode = "all")), doDiag = FALSE),
+    "generalMatrix"
+  ) -
     g[]
   ss <- eigen(A)
 
@@ -497,7 +520,7 @@ test_that("embed_laplacian_matrix -- Undirected, unweighted, DAD case works", {
   g <- sample_gnm(10, 20, directed = FALSE)
 
   no <- 3
-  D12 <- diag(1 / sqrt(degree(g)))
+  D12 <- diag(1 / sqrt(degree(g, mode = "all")))
   A <- D12 %*% g[] %*% D12
   ss <- eigen(A)
 
@@ -596,7 +619,7 @@ test_that("embed_laplacian_matrix -- Undirected, unweighted, I-DAD case works", 
   g <- sample_gnm(10, 20, directed = FALSE)
 
   no <- 3
-  D12 <- diag(1 / sqrt(degree(g)))
+  D12 <- diag(1 / sqrt(degree(g, mode = "all")))
   A <- diag(vcount(g)) - D12 %*% g[] %*% D12
   ss <- eigen(A)
 
@@ -793,7 +816,7 @@ test_that("embed_laplacian_matrix -- Undirected, unweighted, DAD case works", {
   g <- sample_gnm(10, 20, directed = FALSE)
 
   no <- 3
-  D12 <- diag(1 / sqrt(degree(g)))
+  D12 <- diag(1 / sqrt(degree(g, mode = "all")))
   A <- D12 %*% g[] %*% D12
   ss <- eigen(A)
 
@@ -893,7 +916,7 @@ test_that("embed_laplacian_matrix -- Undirected, unweighted, I-DAD case works", 
   g <- sample_gnm(10, 20, directed = FALSE)
 
   no <- 3
-  D12 <- diag(1 / sqrt(degree(g)))
+  D12 <- diag(1 / sqrt(degree(g, mode = "all")))
   A <- diag(vcount(g)) - D12 %*% g[] %*% D12
   ss <- eigen(A)